+++ tipodemovimiento.vb
@@ -1,6 +1,7 @@
 ﻿Imports System.Data.SqlClient
 
 Public Class tipodemovimiento
+    Private editarActivo As Boolean = False
     Sub Agregar()
         On Error GoTo Errores
         If Me.txtNombreMovimiento.Text = "" Or cboxTipoMovimiento.Text = "" Then
@@ -23,27 +24,27 @@ Errores:
 
     Sub Editar()
         On Error GoTo Errores
-        Dim Rs As SqlDataReader
-        Dim itemSeleccionado As String = Me.listTipoMovimiento.SelectedItem.ToString()
 
-        Sql = "SELECT [nom tipomovi], [tip tipomovi] FROM tipomovi WHERE [nom tipomovi] = '" & itemSeleccionado & "'"
-        Instruccion = New SqlCommand(Sql, DaoCon)
-        Rs = Instruccion.ExecuteReader()
+        Dim nombreTipoMovimiento As String = Me.listTipoMovimiento.SelectedItem.ToString()
 
-        If Rs.Read() Then
-            Me.txtNombreMovimiento.Text = Rs("nom tipomovi").ToString()
-            Me.cboxTipoMovimiento.Text = Rs("tip tipomovi").ToString()
+        Dim selectSql As String = "SELECT [id tipomovi], [nom tipomovi], [tip tipomovi] FROM tipomovi WHERE [nom tipomovi] = @nombre"
+        Dim selectCmd As New SqlCommand(selectSql, DaoCon)
+        selectCmd.Parameters.AddWithValue("@nombre", nombreTipoMovimiento)
+        Dim reader As SqlDataReader = selectCmd.ExecuteReader()
+
+        If reader.Read() Then
+            Dim idTipoMovimiento As Integer = reader("id tipomovi").ToString()
+            Dim nomTipoMovimiento As String = reader("nom tipomovi").ToString()
+            Dim tipTipoMovimiento As String = reader("tip tipomovi").ToString()
+
+            Me.txtCodigo.Text = idTipoMovimiento
+            Me.txtNombreMovimiento.Text = nomTipoMovimiento
+            Me.cboxTipoMovimiento.Text = tipTipoMovimiento
+            Me.cboxTipoMovimiento.Enabled = False
         End If
+        reader.Close()
 
-        Rs.Close()
-
-        Sql = "DELETE FROM tipomovi WHERE [nom tipomovi] = '" & itemSeleccionado & "'"
-        Instruccion = New SqlCommand(Sql, DaoCon)
-        Instruccion.ExecuteNonQuery()
-
-        MostrarTipoMovimiento()
         Exit Sub
-
 Errores:
         Select Case Err.Number
             Case Else
@@ -52,17 +53,32 @@ Errores:
         End Select
     End Sub
 
-
     Sub Eliminar()
         If Me.listTipoMovimiento.SelectedItem IsNot Nothing Then
             Dim itemSeleccionado As String = Me.listTipoMovimiento.SelectedItem.ToString()
+
+            Dim countSql As String = "SELECT COUNT(*) FROM movimiento WHERE [ID tipomovi] = (SELECT [ID tipomovi] FROM tipomovi WHERE [nom tipomovi] = @nombre)"
+            Dim countCmd As New SqlCommand(countSql, DaoCon)
+            countCmd.Parameters.AddWithValue("@nombre", itemSeleccionado)
+            Dim count As Integer = CInt(countCmd.ExecuteScalar())
+
+            If count > 0 Then
+                MsgBox("El tipo de movimiento " & itemSeleccionado & " no puede ser eliminado porque está siendo utilizado en un movimiento.", vbExclamation)
+                listTipoMovimiento.ClearSelected()
+                Exit Sub
+            End If
+
             Dim respuesta As MsgBoxResult = MsgBox("¿Está seguro que desea eliminar el movimiento " & itemSeleccionado & "?", vbYesNo + vbQuestion)
             If respuesta = vbYes Then
                 Sql = "DELETE FROM tipomovi WHERE [nom tipomovi] = '" & itemSeleccionado & "'"
                 Instruccion = New SqlCommand(Sql, DaoCon)
                 Instruccion.ExecuteNonQuery()
+            Else
+                respuesta = MsgBoxResult.No
+                Me.listTipoMovimiento.ClearSelected()
             End If
         End If
+
         MostrarTipoMovimiento()
     End Sub
 
@@ -82,15 +98,48 @@ Errores:
         Me.cboxTipoMovimiento.Items.Clear()
         Me.cboxTipoMovimiento.Items.Add("E")
         Me.cboxTipoMovimiento.Items.Add("S")
+        Me.btnEliminar.Enabled = False
+        Me.btnEditar.Enabled = False
         LimpiarCampos(Me)
         MostrarTipoMovimiento()
     End Sub
 
+    Private Sub ActualizarTipoMovimiento()
+        Dim idTipoMovimiento As Integer = Integer.Parse(txtCodigo.Text)
+
+        Dim updateSql As String = "UPDATE tipomovi SET [nom tipomovi] = @nombre WHERE [id tipomovi] = @id"
+        Dim updateCmd As New SqlCommand(updateSql, DaoCon)
+        updateCmd.Parameters.AddWithValue("@nombre", txtNombreMovimiento.Text)
+        updateCmd.Parameters.AddWithValue("@id", idTipoMovimiento)
+        updateCmd.ExecuteNonQuery()
+
+        LimpiarCampos(Me)
+        editarActivo = False
+
+        MostrarTipoMovimiento()
+    End Sub
+
     Private Sub btnAgregar_Click(sender As Object, e As EventArgs) Handles btnAgregar.Click
-        Agregar()
+        If editarActivo = True Then
+            ActualizarTipoMovimiento()
+            Me.cboxTipoMovimiento.Enabled = True
+        Else
+            Agregar()
+        End If
+    End Sub
+
+    Private Sub listTipoMovimiento_SelectedIndexChanged(sender As Object, e As EventArgs) Handles listTipoMovimiento.SelectedIndexChanged
+        If listTipoMovimiento.SelectedIndex >= 0 Then
+            btnEditar.Enabled = True
+            btnEliminar.Enabled = True
+        Else
+            btnEditar.Enabled = False
+            btnEliminar.Enabled = False
+        End If
     End Sub
 
     Private Sub btnEditar_Click(sender As Object, e As EventArgs) Handles btnEditar.Click
+        editarActivo = True
         Editar()
     End Sub
 
@@ -100,5 +149,31 @@ Errores:
 
     Private Sub btnLimpiar_Click(sender As Object, e As EventArgs) Handles btnLimpiar.Click
         LimpiarCampos(Me)
+    End Sub
+
+    Private Sub txtNombreMovimiento_TextChanged(sender As Object, e As EventArgs) Handles txtNombreMovimiento.TextChanged
+        LimitarNombre(txtNombreMovimiento)
+        Me.listTipoMovimiento.ClearSelected()
+    End Sub
+
+    Private Sub cboxTipoMovimiento_SelectedIndexChanged(sender As Object, e As EventArgs) Handles cboxTipoMovimiento.SelectedIndexChanged
+        If listTipoMovimiento.SelectedIndex >= 0 Then
+            btnEditar.Enabled = True
+            btnEliminar.Enabled = True
+        Else
+            btnEditar.Enabled = False
+            btnEliminar.Enabled = False
+        End If
+    End Sub
+
+    Private Sub listTipoMovimiento_DoubleClick(sender As Object, e As EventArgs) Handles listTipoMovimiento.DoubleClick
+        If listTipoMovimiento.SelectedIndex >= 0 Then
+            editarActivo = True
+            Editar()
+        End If
+    End Sub
+
+    Private Sub txtNombreMovimiento_Enter(sender As Object, e As EventArgs) Handles txtNombreMovimiento.Enter
+        listTipoMovimiento.ClearSelected()
     End Sub
 End Class
